Add parameter conflict check and comprehensive tests for text edits
